[Refactor] Notification 모델 수정 #4
