+++ src/models/notification.model.ts
@@ -1,0 +1,7 @@
+export interface INotification {
+    id: number;
+    content: string;
+    diaryId?: number;
+    category: string;
+    date: string;
+}
